pyproject.toml
[project]
name = "status-update-agent-with-dynamic-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[all,http-server]>=0.3.5",
    "fastapi>=0.116.1",
    "fastmcp>=2.12.3",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "agent-operator",
]
